+++ FileShareApi/Controllers/UserController.cs
@@ -4,6 +4,9 @@ using Application.Model;
 using Application.Persistence;
 using Application.Service.EntityMapper;
 using Domain;
+using FileShareApi.Auth;
+using FileShareApi.Auth.Contract;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -22,19 +25,28 @@ namespace FileShareApi.Controllers
         private readonly IEntityMapper _mapper;
         private readonly IUserRepository _userRepository;
         private readonly IEncryptionHelper _encryptionHelper;
-        public UserController(IUserRepository userRepository, IEntityMapper mapper,IEncryptionHelper encryptionHelper)
+        private readonly IAccessControl _accessControl;
+        public UserController(IUserRepository userRepository, IEntityMapper mapper,IEncryptionHelper encryptionHelper,IAccessControl accessControl)
         {
             _userRepository = userRepository;
             _mapper = mapper;
             _encryptionHelper = encryptionHelper;
+            _accessControl = accessControl;
         }
         [HttpPost]
         public async Task<IHttpActionResult> Post([FromBody] CreateUserDTO user)
         {
             try
             {
-                user.Password = _encryptionHelper.EncryptString(user.Password);
-                return Ok(await _userRepository.Add(_mapper.EntityMap<Domain.User>(user)));
+                if (_accessControl.Check(Request))
+                {
+                    user.Password = _encryptionHelper.EncryptString(user.Password);
+                    return Ok(await _userRepository.Add(_mapper.EntityMap<Domain.User>(user)));
+                }
+                else
+                {
+                    return Unauthorized();
+                }
             }
             catch (Exception ex)
             {
@@ -46,17 +58,22 @@ namespace FileShareApi.Controllers
         {
             try
             {
-                var Currentuser = _userRepository.GetUserById(user.Id);
-                if (Currentuser == null) return NotFound();
-                else
+                if (_accessControl.Check(Request))
                 {
-                    user.Password = _encryptionHelper.EncryptString(user.Password);
-                    Currentuser = _mapper.EntityMap<User>(user, Currentuser);
-                    if (await _userRepository.Update(Currentuser))
-                        return Ok(true);
+                    var Currentuser = _userRepository.GetUserById(user.Id);
+                    if (Currentuser == null) return NotFound();
                     else
-                        return InternalServerError();
+                    {
+                        user.Password = _encryptionHelper.EncryptString(user.Password);
+                        Currentuser = _mapper.EntityMap<User>(user, Currentuser);
+                        if (await _userRepository.Update(Currentuser))
+                            return Ok(true);
+                        else
+                            return InternalServerError();
+                    }
                 }
+                else
+                    return Unauthorized();
             }
             catch (Exception ex)
             {
@@ -68,15 +85,20 @@ namespace FileShareApi.Controllers
         {
             try
             {
-                var Currentuser = _userRepository.GetUserById(id);
-                if (Currentuser == null) return NotFound();
-                else
+                if (_accessControl.Check(Request))
                 {
-                    if (await _userRepository.Delete(Currentuser))
-                        return Ok(true);
+                    var Currentuser = _userRepository.GetUserById(id);
+                    if (Currentuser == null) return NotFound();
                     else
-                        return InternalServerError();
+                    {
+                        if (await _userRepository.Delete(Currentuser))
+                            return Ok(true);
+                        else
+                            return InternalServerError();
+                    }
                 }
+                else
+                    return Unauthorized();
             }
             catch (Exception ex)
             {
@@ -88,7 +110,10 @@ namespace FileShareApi.Controllers
         {
             try
             {
-                return Ok(_mapper.EntityMap<List<UserDTO>>(_userRepository.GetUsers(IncludeDisabled, Skip, PageSize).ToList()));
+                if (_accessControl.Check(Request))
+                    return Ok(_mapper.EntityMap<List<UserDTO>>(_userRepository.GetUsers(IncludeDisabled, Skip, PageSize).ToList()));
+                else
+                    return Unauthorized();
             }
             catch (Exception ex)
             {
@@ -100,12 +125,16 @@ namespace FileShareApi.Controllers
         {
             try
             {
-                var result = new UserDTO();
-                var user = _userRepository.GetUserById(id);
-                user.Password = _encryptionHelper.DecryptString(user.Password);
-                user.Password = Convert.ToBase64String(Encoding.UTF8.GetBytes(user.Password));
-                result = _mapper.EntityMap<UserDTO>(user);
-                return Ok(result);
+                if (_accessControl.Check(Request))
+                {
+                    var result = new UserDTO();
+                    var user = _userRepository.GetUserById(id);
+                    result = _mapper.EntityMap<UserDTO>(user);
+                    result.Password = _encryptionHelper.LocalRSAEncrypt(_encryptionHelper.DecryptString(user.Password));
+                    return Ok(result);
+                }
+                else
+                    return Unauthorized();
             }
             catch (Exception ex)
             {
@@ -119,7 +148,7 @@ namespace FileShareApi.Controllers
             {
                 user.Password = _encryptionHelper.EncryptString(user.Password);
                 var loginResult = _userRepository.LoginUser(user.Username, user.Password);
-                LoginResult result = new LoginResult() {  successLogin = false, User = null };
+                LoginResult result = new LoginResult() { successLogin = false, User = null };
                 switch (loginResult.Item1)
                 {
                     case 1:
